Fixed bugs in pyDEV_INFO and correctly implemented pyDEV_CALLBACKS

+++ libnfc/chip/chip.py
@@ -1,6 +1,6 @@
-from ctypes import c_int
+from ctypes import c_ubyte
 
 # chip type
-CT_PN531                    = c_int(0x10)
-CT_PN532                    = c_int(0x20)
-CT_PN533                    = c_int(0x30)
+CT_PN531                    = c_ubyte(0x10)
+CT_PN532                    = c_ubyte(0x20)
+CT_PN533                    = c_ubyte(0x30)
